+++ server/src/server.ts
@@ -72,15 +72,61 @@ import {
 import { getFoldingRanges } from "./foldingRanges";
 import { typeDefinition } from "./typeDefinition";
 import { resolve } from "path";
+import { FileWatcher } from "./fileWatcher";
 
-let contextAware: ContextAware[] = [];
+const contextAware: ContextAware[] = [];
 let activeContext: Promise<ContextAware | undefined>;
 let activeFileUri: string | undefined;
-
 let debounce = new WeakMap<
   ContextAware,
   { abort: AbortController; promise: Promise<void> }
 >();
+const fileWatchers = new Map<string, FileWatcher>();
+
+const addContext = (context: ContextAware) => {
+  if (contextAware.some((ctx) => ctx === context)) {
+    return;
+  }
+
+  contextAware.push(context);
+  watchContextFiles(context);
+};
+
+const watchContextFiles = async (context: ContextAware) => {
+  await context.stable();
+  context.getContextFiles().forEach((file) => {
+    if (file.endsWith(".h")) return;
+    if (!fileWatchers.has(file)) {
+      fileWatchers.set(
+        file,
+        new FileWatcher(
+          file,
+          onChange,
+          (file) => !!documents.get(`file://${file}`)
+        )
+      );
+    }
+    fileWatchers.get(file)?.watch();
+  });
+};
+
+const deleteContext = async (context: ContextAware) => {
+  const index = contextAware.indexOf(context);
+  if (index === -1) {
+    return;
+  }
+
+  contextAware.splice(index, 1);
+
+  unwatchContextFiles(context);
+};
+
+const unwatchContextFiles = async (context: ContextAware) => {
+  await context.stable();
+  context
+    .getContextFiles()
+    .forEach((file) => fileWatchers.get(file)?.unwatch());
+};
 
 const isStable = (context: ContextAware) => {
   const d = debounce.get(context);
@@ -188,7 +234,8 @@ const cleanUpAdHocContext = async (context: ContextAware) => {
         `cleaning up context with ID ${c.name} and uri ${c.parser.uri}`
       );
     });
-    contextAware = contextAware.filter((c) => contextToClean.indexOf(c) === -1);
+
+    contextToClean.forEach(deleteContext);
   }
 };
 
@@ -380,7 +427,7 @@ connection.onDidChangeConfiguration((change) => {
 
   let adhocContexts = getAdhocContexts(oldSettngs);
 
-  contextAware = [];
+  contextAware.forEach(deleteContext);
 
   debounce = new WeakMap<
     ContextAware,
@@ -407,7 +454,7 @@ connection.onDidChangeConfiguration((change) => {
       context.overlays,
       context.ctxName
     );
-    contextAware.push(newContext);
+    addContext(newContext);
     console.log(
       `New context with ID ${newContext.name} for ${context.dtsFile}`
     );
@@ -416,9 +463,8 @@ connection.onDidChangeConfiguration((change) => {
   // resolve global with cwd
   const resolvedGlobal = resolveGlobal();
 
-  contextAware = [
-    ...contextAware,
-    ...adhocContexts.map((c) => {
+  adhocContexts
+    .map((c) => {
       const bindingType = globalSettings.defaultBindingType;
       const context = new ContextAware(
         c.parser.uri,
@@ -438,8 +484,8 @@ connection.onDidChangeConfiguration((change) => {
       );
       console.log(`New context with ID ${context.name} for ${c.parser.uri}`);
       return context;
-    }),
-  ];
+    })
+    .forEach(addContext);
 
   adhocContexts = getAdhocContexts(globalSettings);
   adhocContexts.forEach(cleanUpAdHocContext);
@@ -763,18 +809,7 @@ const standardTypeToLinkedMessage = (issue: StandardTypeIssue) => {
   }
 };
 
-// // The content of a text document has changed. This event is emitted
-// // when the text document first opened or when its content has changed.
-documents.onDidChangeContent(async (change) => {
-  console.log("Content changed");
-  const uri = change.document.uri.replace("file://", "");
-
-  const text = change.document.getText();
-  const tokenProvider = getTokenizedDocumentProvider();
-  if (!tokenProvider.needsRenew(uri, text)) return;
-
-  tokenProvider.renewLexer(uri, text);
-
+const onChange = async (uri: string) => {
   await initialSettingsProvided;
 
   const contexts = await findContexts(contextAware, uri);
@@ -803,7 +838,7 @@ documents.onDidChangeContent(async (change) => {
         : undefined
     );
     console.log(`New ad hoc context with ID ${newContext.name} for ${uri}`);
-    contextAware.push(newContext);
+    addContext(newContext);
     updateActiveContext(uri);
     await newContext.parser.stable;
     cleanUpAdHocContext(newContext);
@@ -821,7 +856,9 @@ documents.onDidChangeContent(async (change) => {
           const itemsToClear = generateClearWorkspaceDiagnostics(
             context.context
           );
+          unwatchContextFiles(context.context);
           await context.context.reevaluate(uri);
+          watchContextFiles(context.context);
           clearWorkspaceDiagnostics(context.context, itemsToClear);
           reportWorkspaceDiagnostics(context.context).then((d) => {
             d.items
@@ -845,6 +882,20 @@ documents.onDidChangeContent(async (change) => {
       debounce.set(context.context, { abort, promise });
     });
   }
+};
+
+// // The content of a text document has changed. This event is emitted
+// // when the text document first opened or when its content has changed.
+documents.onDidChangeContent(async (change) => {
+  const uri = change.document.uri.replace("file://", "");
+
+  const text = change.document.getText();
+  const tokenProvider = getTokenizedDocumentProvider();
+  if (!tokenProvider.needsRenew(uri, text)) return;
+
+  console.log("Content changed");
+  tokenProvider.renewLexer(uri, text);
+  onChange(uri);
 });
 
 const generateClearWorkspaceDiagnostics = (context: ContextAware) =>