+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # 0.1.1 - UNRELEASED
 
+### Added
+
+- File watchers for files used by contexts.
+
 ### Fixed
 
 - Diagnostics when property is missing ';'. Now code does not assume ',' is missing if token is at end of line.